Use PAGE_SAVE_AS and PAGE_URL to determine where pages end up
This commit stops using save_as in individual markdown files in favour
of giving them a "slug", which is used to construct both the URL
generated for that file by pelican and also the location to save the
rendered output.

This allows using links generated from relative paths to source files in
place of links directly to output html.