Replace all internal links to reference relative source paths
This commit replaces all the intra-site links with relative paths to the
relevant source files. This ensures that the correct link is generated
in the output HTML (rather than articles in subdirectories having either
hardcoded absolute or relative links).

It also allows build-time checking for dead internal links, since this
method gives a warning on build if the target source file isn't found.