Update for final beta stage
BuildStream 1.95.2 is a feature-complete beta release which is ready for
wide testing.

The only remaining blocker to a BuildStream 2.0 release is completing
some administrative activities related to becoming a full Apache
Foundation project, which are expected to complete in late September.

This commit should be reverted once the final 2.0 release happens.